+++ src/Common/Settings.cs
@@ -18,7 +18,8 @@
             var solarAnalysisOptionsViewModel = new SolarAnalysis.ViewModels.SolarAnalysisOptonsViewModel();
             var spellCheckerOptionsViewModel = new SpellChecker.ViewModels.SpellCheckerOptionsViewModel();
             var sheetCopierViewModel = new SheetCopier.ViewModels.SheetCopierSettingsViewModel();
-            var vm = new ViewModels.SettingsViewModel(incrementViewModel, roomConverterViewModel, viewUtilitiesViewModel,solarAnalysisOptionsViewModel, sheetCopierViewModel,spellCheckerOptionsViewModel);
+            var runScriptSettingViewModel = new RunScript.ViewModels.RunScriptSettingsViewModel();
+            var vm = new ViewModels.SettingsViewModel(incrementViewModel, roomConverterViewModel, viewUtilitiesViewModel,solarAnalysisOptionsViewModel, sheetCopierViewModel,spellCheckerOptionsViewModel, runScriptSettingViewModel);
             SCaddinsApp.WindowManager.ShowDialogAsync(vm, null, ViewModels.SettingsViewModel.DefaultWindowSettings);
             return Result.Succeeded;
         }